Move check_submit_date.py to common
