+++ cs251tk/common/check_submit_date.py
@@ -1,6 +1,6 @@
 import os
 from dateutil.parser import parse
-from ...common import run, chdir
+from cs251tk.common import run, chdir
 
 
 def check_dates(spec_id, username, spec, basedir):
